+++ xlsxreport/__init__.py
@@ -1,15 +1,12 @@
-#!/usr/bin/env python
-# -*- coding: utf-8 -*-
-
 """
 XlsxReport is a Python package to automatically generate formatted excel reports
-from quantitative mass spectrometry result tables. YAML config files are used to
-describe the content of a result table and the format of the excel report.
+from quantitative mass spectrometry result tables. YAML template files are used to
+describe how the content of a result table should be formatted in the Excel report.
 """
 
 
 from .writer import Reportbook
-from .appdir import locate_data_dir, get_config_file, setup_data_dir
+from .appdir import locate_data_dir, get_template_file, setup_data_dir
 
 
 __author__ = "David M. Hollenstein"