+++ xlsxreport/scripts/setup_appdata_dir.py
@@ -13,10 +13,16 @@ def cli() -> None:
     data_dir = xlsxreport.locate_data_dir()
     if os.path.isdir(data_dir):
         click.echo(f"App data directory for XlsxReport found at:")
-        click.echo(f"\t{data_dir}")
+        click.echo(f"  {data_dir}")
     else:
         click.echo(f"Creating XlsxReport folder in the local user data directory at:")
-        click.echo(f"\t{data_dir}")
-    click.echo("Copying default XlsxReport config files to the app data directory ...")
+        click.echo(f"  {data_dir}")
+    click.echo(
+        "Copying default XlsxReport template files to the app data directory ..."
+    )
     xlsxreport.setup_data_dir()
-    click.echo(f"\tConfig files were successfully copied.")
+    click.echo(f"  Template files were successfully copied.")
+
+
+if __name__ == "__main__":
+    cli()