+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ## Upcoming release
 
 ### Changed
+- Replaced the term "config" with "template" throughout the codebase.
 - Changed default value of `evaluate_log2_transformation` to False.
 - Changed default value of `write_supheader` to False.
 - Supheader cells are now still merged when no supheader name is specified.